Added a CoordinateData class for the Data abstract class to manage incoming coordinate info. It makes each set of three coordinates into a coordinate object, and adds them to a list of such objects which are stored as a variable of the data class. The data class now has an appropriate setter to do this.
That list of coordinate objects will then be used for
calorie calculation. Similarly, the next thing to do will be to store
time values as temporal objects in order to calculate time differences
between coordinate values, for the purpose of calorie calculation.
To do this, a list of time values will be passed
into the Data object.

This information will then be compared with the heart rate during that
period to estimate calories burned.

When there are times that are to difficult to calculate the calories
using the coordinates, an attempt will be made to give an
estimation purely based on the heart rate over that time period.