Finished the calorie calculations for everything. This included making the
calorieCalculationSetup

function for the calories from walking,
running, and cycling. That function is used by

calculateCaloriesFromStats

to calculate the calories burned whilst performing those types of
exercise, in which speed is a useful factor for determining calories
burned. To that end,

calculateSpeedsBetweenPointsKph and Mph variant

were created. Also created was a method to

calculateGradientsBetweenPoints

, and calculations to vary the MET value according to the
speed were included. For the sports in which basing the calorie
calculations on speed and MET values over time was not feasible due to
the specificity of certain activities (e.g. swimming has different MET
values for each type of stroke, and rock climbing has only general MET
values)

calculateCaloriesFromUserStatsAndHeartRateAndTime

was created. This can be a fallback function to use if anything goes
awry with the calorie calculating function for running, walking, and
biking also (Though the function they have should be more accurate).
This fallback
function could have been used from the beginning to save time, but I did
not find
out the formula for it until near the very end(nor did I know this
general formula existed). However,
the extra statistics I have calculated for the data class for the
purpose of doing the running walking biking calorie calculations will be
very useful for making
visual graph representations of certain stats, making Joels job easier
anyway. So the time was still mostly well spent.